+++ composeApp/src/commonMain/kotlin/scenes/mainscreen/MainScreen.kt
@@ -1,5 +1,6 @@
 package scenes.mainscreen
 
+import sharedData.AppColors
 import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -28,7 +29,6 @@ import cafe.adriel.voyager.navigator.tab.TabNavigator
 import kotlinproject.composeapp.generated.resources.Res
 import kotlinproject.composeapp.generated.resources.manrope_regular
 import org.jetbrains.compose.resources.Font
-import sharedData.AppColors
 import sharedData.getScreenWidth
 
 class MainScreen : Screen {
