+++ composeApp/src/commonMain/kotlin/scenes/onboarding/OnboardingScreen.kt
@@ -1,5 +1,6 @@
 package scenes.onboarding
 
+import sharedData.AppColors
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -46,7 +47,6 @@ import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinViewModel
 import scenes.signinscreen.SignInScreen
-import sharedData.AppColors
 
 class OnboardingScreen : Screen {
     @Composable
