+++ composeApp/src/commonMain/kotlin/scenes/profilescreen/ProfileScreen.kt
@@ -1,5 +1,6 @@
 package scenes.profilescreen
 
+import sharedData.AppColors
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -37,9 +38,8 @@ import kotlinproject.composeapp.generated.resources.version
 import org.jetbrains.compose.resources.Font
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinViewModel
-import sharedData.AppColors
 import sharedData.getAppVersion
-import ui.screenContent.CommonBackground
+import ui.CommonBackground
 
 
 class ProfileScreen : Screen {
