+++ composeApp/src/commonMain/kotlin/scenes/signinscreen/SignInScreen.kt
@@ -1,5 +1,6 @@
 package scenes.signinscreen
 
+import sharedData.AppColors
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
@@ -55,7 +56,6 @@ import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinViewModel
 import scenes.mainscreen.MainScreen
-import sharedData.AppColors
 
 class SignInScreen : Screen {
     @Composable
